Update ilyushin.xml
I have added three versions of IL-18 and two versions of IL-62. I used Russian and Czech wikipedias as main sources.
It is a real challenge to assign passenger numbers and range, because the range strongly depends on configuartions. For example the IL-18D could either fly 6500 km with 65 passengers, or 5000 km with 120. The same applies ot IL-62: with 100 passengers it flew 10000km. So I took averages for the IL-18s, and max. take-off weight for the IL-62. 
The color of the text reversed in the code after the IL-62, I could not find a reason for it.